requirements.txt
pyaudio
websocket-client
google-generativeai 
numpy
PyQt6
huggingface-hub
pyyaml
python-dotenv